tests/wpt/web-platform-tests/html/canvas/element/text-styles/2d.text.font.parse.system.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.text.font.parse.system</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.font.parse.system</h1>
<p class="desc">System fonts must be computed to explicit values</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("System fonts must be computed to explicit values");
_addTest(function(canvas, ctx) {

ctx.font = 'message-box';
_assertDifferent(ctx.font, 'message-box', "ctx.font", "'message-box'");


});
</script>
